move preload_pages to rendering.py
